more attemps to get config-comp working
